+++ src/routes/roro/TaskDetails/TaskSummary.jsx
@@ -2,10 +2,10 @@ import React from 'react';
 import dayjs from 'dayjs';
 import utc from 'dayjs/plugin/utc';
 import relativeTime from 'dayjs/plugin/relativeTime';
-import { LONG_DATE_FORMAT, RORO_TOURIST, RORO_TOURIST_SINGLE_ICON, RORO_TOURIST_GROUP_ICON } from '../../constants';
-import getMovementModeIcon from '../../utils/getVehicleModeIcon';
-import { modifyRoRoPassengersTaskList, hasVehicle, hasTrailer, hasDriver } from '../../utils/roroDataUtil';
-import { formatMovementModeIconText } from '../../utils/stringConversion';
+import { LONG_DATE_FORMAT, RORO_TOURIST, RORO_TOURIST_SINGLE_ICON, RORO_TOURIST_GROUP_ICON } from '../../../constants';
+import getMovementModeIcon from '../../../utils/getVehicleModeIcon';
+import { modifyRoRoPassengersTaskList, hasVehicle, hasTrailer, hasDriver } from '../../../utils/roroDataUtil';
+import { formatMovementModeIconText } from '../../../utils/stringConversion';
 
 import '../__assets__/TaskDetailsPage.scss';
