+++ src/routes/roro/TaskDetails/TaskVersionsMode/RoRoAccompaniedMode.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { calculateTaskVersionTotalRiskScore } from '../../../utils/rickScoreCalculator';
+import { calculateTaskVersionTotalRiskScore } from '../../../../utils/rickScoreCalculator';
 
 import {
   renderTargetingIndicatorsSection,
@@ -19,7 +19,7 @@ import {
   extractTaskVersionsBookingField,
   modifyRoRoPassengersTaskList,
   modifyCountryCodeIfPresent,
-} from '../../../utils/roroDataUtil';
+} from '../../../../utils/roroDataUtil';
 
 const renderFirstColumn = (version, movementMode) => {
   const targIndicatorsField = version.find(({ propName }) => propName === 'targetingIndicators');
