+++ src/routes/roro/TaskDetails/TaskVersionsMode/SelectorMatchesTaskVersion.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Tab, TabList, Tabs, TabPanel } from 'react-tabs';
 import { v4 as uuidv4 } from 'uuid';
-import '../../../__assets__/ReactTabs.scss';
+import '../../../../__assets__/ReactTabs.scss';
 
 const warningCodesMapping = {
   VIOL: 'Violence',
