+++ src/routes/roro/__fixtures__/variableInstanceStatusComplete.fixture.json
