+++ src/routes/roro/__fixtures__/variableInstanceStatusNew.fixture.json
